requirements.txt
#Object Detection


tensorflow
opencv-python
numpy
pillow
